backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.1.0
celery[redis]==5.4.0
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
torch==2.4.1
transformers==4.44.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
